[Feat] 아이템, 전시 조회시 이미지 추가

+++ src/main/java/com/fisa/wooriarte/spaceItem/domain/SpaceItem.java
@@ -55,7 +55,7 @@ public class SpaceItem {
     private Boolean parking;
 
     @Column
-    private int fee;
+    private Long fee;
 
     @Column
     private String phone;
